Extract AbstractArticleCrudServiceImpl without crepo
